kafka_producer/requirements.txt
kafka-python
faker
numpy
pandas
